requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.4
cryptography==43.0.3
dependency-injector==4.41.0
dnspython==2.7.0
email-validator==2.1.0.post1
exceptiongroup==1.2.2
fastapi==0.115.4
fastapi-crudrouter==0.7.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jwt==1.3.1
packaging==24.2
passlib[bcrypt]==1.7.4
paho-mqtt==1.6.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==1.10.14
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==6.0.0
PyYAML==6.0.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tomli==2.1.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
python-dateutil>=2.8.2
